API connection - Creating transaction context
