+++ src/App.tsx
@@ -1,6 +1,7 @@
 import { ThemeProvider } from "styled-components";
 
 import { Transactions } from "./pages/Transactions";
+import { TransactionsProvider } from "./contexts/TransactionsContext";
 
 import { defaultTheme } from "./styles/themes/default";
 import { GlobalStyle } from "./styles/global";
@@ -12,7 +13,9 @@ export function App() {
     <ThemeProvider theme={defaultTheme}>
       <GlobalStyle />
 
-      <Transactions />
+      <TransactionsProvider>
+        <Transactions />
+      </TransactionsProvider>
     </ThemeProvider>
   )
 }
